01-PROJECT-FLASHCARDS/backend/requirements.txt
fastapi
uvicorn
sqlmodel
python-multipart
pypdf
google-generativeai
python-dotenv
pytest
httpx
pytest-asyncio
fastmcp
mcp
pydantic-ai
nest-asyncio
